reorganize realms
