+++ lib/templates/gametemplate.js
@@ -2873,7 +2873,7 @@ class GameTemplate extends ModTemplate {
         // TODO : change id to unique identifier (indexed on card array / which might be hand)
         // move deck's card id into data-id slot and then update ALL the games!
         //
-console.log(cardarray[i] + " --- " + deckid);
+        //console.log(cardarray[i] + " --- " + deckid);
 
         html += `<div id="${cardarray[i]}" class="card hud-card ${cardarray[i]}">${this.returnCardImage(
           cardarray[i],
