--- mods/realms/compile
@@ -1,27 +1,0 @@
-
-rm ../../web/saito/saito.js
-
-#
-# class constructor
-#
-cat ./src/realms-init.js > ./realms.js
-
-#
-# eventually every card compiles in separately, for now 
-# all decks and deck-management provided by realms-cards
-#
-#for file in ./src/cards/core/*
-#do
-#  cat $file >> ./realms.js
-#done
-
-cat ./src/realms-cards.js >> ./realms.js
-cat ./src/realms-gameloop.js >> ./realms.js
-cat ./src/realms-state.js >> ./realms.js
-cat ./src/realms-events.js >> ./realms.js
-cat ./src/realms-player.js >> ./realms.js
-cat ./src/realms-display.js >> ./realms.js
-cat ./src/realms-end.js >> ./realms.js
-
-npm run compile $1
-
