--- mods/realms/src/realms-end.js
@@ -1,5 +1,0 @@
-
-}
-
-module.exports = Realms;
-
