+++ mods/realms/web/index.html
@@ -14,7 +14,6 @@
     <script type="text/javascript" src="/saito/lib/jquery/jquery-ui.min.js"></script>
 
     <link rel="stylesheet" href="/saito/lib/font-awesome-6/css/all.css" type="text/css" media="screen" />
-    <link rel="manifest" href="/manifest.json">
 
     <link rel="stylesheet" type="text/css" href="/saito/saito.css" />
     <link rel="stylesheet" type="text/css" href="/saito/game.css" />
